+++ docs/CHANGELOG.md
@@ -1,1 +1,6 @@
+# Changelog
 
+## 0-α1
+- Created the project in Github
+- Updated all code to contain copyright and licensing
+- Added documentation